third_party/WebKit/LayoutTests/fast/dom/Window/window-scaled-viewport-properties-expected.txt

This test makes sure the window properties related to the viewport remain correct under pinch-to-zoom.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


===Unscaled===

PASS window.innerWidth is 800
PASS window.innerHeight is 600

===Pinch Zoom in to 2X===

PASS window.innerWidth is 400
PASS window.innerHeight is 300
PASS window.scrollX is 0
PASS window.scrollY is 0
PASS window.scrollX is 10
PASS window.scrollY is 20
PASS window.scrollX is 1600
PASS window.scrollY is 1200
PASS window.scrollX is 0
PASS window.scrollY is 0
PASS window.scrollX is 1600
PASS window.scrollY is 1200
PASS window.scrollX is 0
PASS window.scrollY is 0

===Pinch Out to 0.5X===

PASS window.innerWidth is 1600
PASS window.innerHeight is 1200
PASS window.scrollX is 0
PASS window.scrollY is 0
PASS window.scrollX is 10
PASS window.scrollY is 20
PASS window.scrollX is 400
PASS window.scrollY is 300
PASS window.scrollX is 0
PASS window.scrollY is 0
PASS window.scrollX is 400
PASS window.scrollY is 300
PASS window.scrollX is 0
PASS window.scrollY is 0

===Test OnScroll===

PASS window.innerWidth is 800
PASS window.innerHeight is 600
PASS window.scrollX is 0
PASS window.scrollY is 0
PASS OnScroll called for scroll #1
PASS OnScroll called for scroll #2
PASS OnScroll called for scroll #3
PASS OnScroll called for scroll #4

PASS successfullyParsed is true

TEST COMPLETE
